Fix paste from clipboard
