+++ src/main/deltachat/extras.ts
@@ -1,4 +1,4 @@
-import { app as rawApp } from 'electron'
+import { app as rawApp, clipboard } from 'electron'
 import { ExtendedAppMainProcess } from '../types'
 import { getLogger } from '../../shared/logger'
 import SplitOut from './splitout'
@@ -9,6 +9,9 @@ import { txCoreStrings } from './login'
 const app = rawApp as ExtendedAppMainProcess
 const log = getLogger('main/deltachat/extras')
 import { refresh as refreshMenu } from '../menu'
+import { join } from 'path'
+import mimeTypes from 'mime-types'
+import { writeFile } from 'fs/promises'
 
 // Extras, mainly Electron functions
 export default class Extras extends SplitOut {
@@ -46,5 +49,21 @@ export default class Extras extends SplitOut {
 
   async getAvailableThemes() {
     return await getAvailableThemes()
+  }
+
+  async writeClipboardToTempFile() {
+    const formats = clipboard.availableFormats().sort()
+    log.debug('Clipboard available formats:', formats)
+    if (formats.length <= 0) {
+      throw new Error('No files to write')
+    }
+    const pathToFile = join(
+      rawApp.getPath('temp'),
+      `paste.${mimeTypes.extension(formats[0]) || 'bin'}`
+    )
+    const buf = clipboard.readBuffer(formats[0])
+    log.debug(`Writing clipboard ${formats[0]} to file ${pathToFile}`)
+    await writeFile(pathToFile, buf, 'binary')
+    return pathToFile
   }
 }
